@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import org.apache.commons.lang3.ArchUtils;
+
 import org.sng.main.util.KeyWord;
 
 public class InputData {
@@ -27,8 +27,11 @@ public class InputData {
 
     public enum NetworkType{
         IPMETRO("ipmetro"),
+        IPMETRO_NEW("ipmetro_new"),
         IPRAN("ipran"),
-        CLOUDNET("cloudnet");
+        IPRAN_NEW("ipran_new"),
+        CLOUDNET("cloudnet"),
+        CLOUDNET_NEW("cloudnet_new");
 
         private String name;
 
@@ -38,6 +41,30 @@ public class InputData {
 
         public String getName() {
             return name;
+        }
+
+        public static NetworkType getType(String name) {
+            switch (name.toLowerCase()) {
+                case "ipmetro":{
+                    return IPMETRO;
+                }
+                case "ipmetro_new":{
+                    return IPMETRO_NEW;
+                }
+                case "ipran":{
+                    return IPRAN;
+                }
+                case "ipran_new":{
+                    return IPRAN_NEW;
+                }
+                case "cloudnet":{
+                    return CLOUDNET;
+                }
+                case "cloudnet_new":{
+                    return CLOUDNET_NEW;
+                }
+            }
+            return IPMETRO;
         }
     }
 
@@ -50,6 +77,7 @@ public class InputData {
     private static List<String> ipmetroCaseType1 = new ArrayList<>(Arrays.asList("1.1", "1.2", "1.3", "2.1", "2.2", "4.1"));
     private static List<String> ipmetroCaseType2 = new ArrayList<>(Arrays.asList("2.3", "3.1"));
     private static List<String> ipranCaseType1 = new ArrayList<>(Arrays.asList("1.1", "1.2", "1.3", "2.1", "2.2", "2.4", "2.5"));
+    private static List<String> cloudnetCaseType1 = new ArrayList<>(Arrays.asList("1.1", "1.2", "1.3", "1.4", "2.1", "2.2", "3.1", "4.1"));
 
     
     private static Map<String, String> errIpmetroDstNameMap = new HashMap<>();
@@ -58,6 +86,8 @@ public class InputData {
     private static Map<String, String> corIpmetroDstIpMap = new HashMap<>();
     private static Map<String, String> errIpranDstNameMap = new HashMap<>();
     private static Map<String, String> errIpranDstIpMap = new HashMap<>();
+    private static Map<String, String> errCloudnetDstNameMap = new HashMap<>();
+    private static Map<String, String> errCloudnetDstIpMap = new HashMap<>();
 
     private static String relativeErrProvRootPath = "networks/provenanceInfo/";
     private static String relativePeerInfoRootPath = "networks/peerInfo/";
@@ -68,7 +98,7 @@ public class InputData {
     private static String relativeIgpResultRootPath = "igp_reqs/";
     private static String relativeSseProvRootPath = "sse_provenanceInfo/";
 
-    //------------------------------------IPMETRO---------------------------------------------------
+    //------------------------------------IPMETRO[ERROR]---------------------------------------------------
     private static String errIpmetroDstName1 = "BNG30";
     private static String errIpmetroDstIp1 = "179.0.0.117/30";
 
@@ -76,23 +106,59 @@ public class InputData {
     private static String errIpmetroDstIp2 = "209.0.0.12/30";
 
     private static Set<String> errIpmetroSrcNameList1 = new HashSet<>(Arrays.asList("BNG1"));
-    //------------------------------------IPMETRO---------------------------------------------------
+    //------------------------------------IPMETRO[ERROR]---------------------------------------------------
 
-    //----------------------------------------------------------------------------------------
+    //------------------------------------IPMETRO[CORRECT]----------------------------------------------------
     private static String corIpmetroDstName1 = "BNG3";
     private static String corIpmetroDstIp1 = "179.0.0.9/30";
 
     private static String corIpmetroDstName2 = "BR3";
     private static String corIpmetroDstIp2 = "209.0.0.9/30";
-    //----------------------------------------------------------------------------------------
+    //-------------------------------------IPMETRO[CORRECT]---------------------------------------------------
 
-    //------------------------------------IPRAN----------------------------------------------------
+    //------------------------------------IPMETRO_NEW[ERROR]---------------------------------------------------
+    private static List<String> ipmetroNewCaseType1 = new ArrayList<>(Arrays.asList("1.1", "1.2", "1.3", "2.1", "3.1"));
+    private static List<String> ipmetroNewCaseType2 = new ArrayList<>(Arrays.asList("2.2"));
+    private static List<String> ipmetroNewCaseType3 = new ArrayList<>(Arrays.asList("4.1"));
+
+    private static String errIpmetroNewDstName1 = "BNG20";
+    private static String errIpmetroNewDstIp1 = "173.0.0.77/32";
+
+    private static Set<String> errIpmetroNewSrcNameList1 = new HashSet<>(Arrays.asList("BNG10"));
+
+    private static String errIpmetroNewDstName2 = "BR2";
+    private static String errIpmetroNewDstIp2 = "203.0.0.5/32";
+
+    private static String errIpmetroNewDstIp3 = "20.20.20.20/32";
+    //------------------------------------IPMETRO_NEW[ERROR]---------------------------------------------------
+
+    //------------------------------------IPRAN[CORRECT]----------------------------------------------------
     private static String errIpranDstName1 = "CSG1-1-1";
     private static String errIpranDstIp1 = "191.0.0.0/30";
 
     private static Set<String> errIpranSrcNameList1 = new HashSet<>(Arrays.asList("RSG1"));
-    //------------------------------------IPRAN---------------------------------------------------
-    
+    //------------------------------------IPRAN[CORRECT]---------------------------------------------------
+
+    //------------------------------------IPRAN_NEW[ERROR]----------------------------------------------------
+    private static String errIpranNewDstName1 = "RSG3";
+    private static String errIpranNewDstIp1 = "183.0.0.17/32";
+
+    private static Set<String> errIpranNewSrcNameList1 = new HashSet<>(Arrays.asList("CSG1-1-1"));
+    //------------------------------------IPRAN_NEW[ERROR]---------------------------------------------------
+
+    //------------------------------------CLOUDNET[ERROR]---------------------------------------------------
+    private static String errCloudnetDstName1 = "cloudPE-4";
+    private static String errCloudnetDstIp1 = "50.0.0.13/32";
+
+    private static Set<String> errCloudnetSrcNameList1 = new HashSet<>(Arrays.asList("U1-1-1-1"));
+    //------------------------------------CLOUDNET[ERROR]---------------------------------------------------
+
+    //------------------------------------CLOUDNET[CORRECT]---------------------------------------------------
+    private static String corCloudnetDstName1 = "cloudPE-1";
+    private static String corCloudnetDstIp1 = "90.0.0.1/32";
+
+    private static Set<String> corCloudnetSrcNameList1 = new HashSet<>(Arrays.asList("U1-1-1-1"));
+    //------------------------------------CLOUDNET[CORRECT]---------------------------------------------------
     
 
     public InputData() {
@@ -127,8 +193,14 @@ public class InputData {
             case IPRAN: {
                 return errIpranSrcNameList1;
             }
+            case IPRAN_NEW:{
+                return errIpranNewSrcNameList1;
+            }
             case CLOUDNET: {
-                return new HashSet<String>();
+                return errCloudnetSrcNameList1;
+            }
+            case IPMETRO_NEW:{
+                return errIpmetroNewSrcNameList1;
             }
         }
         throw new IllegalArgumentException("Invalid network Type!");
@@ -145,89 +217,105 @@ public class InputData {
 
     // 待生成/写入的文件
     public String getIgpRequirementFilePath(String keyString, NetworkType type) {
-        String relativePath = concatFilePath(relativeIgpResultRootPath, concatFilePath(type.name(), "case" + keyString +".json"));
+        String relativePath = concatFilePath(relativeIgpResultRootPath, concatFilePath(type.getName(), "case" + keyString +".json"));
         String filePath = concatFilePath(projectRootPath, relativePath);
         return filePath;
     }
 
     // 待生成/写入的文件
     public String getConditionFilePath(String keyString, NetworkType type) {
-        String relativePath = concatFilePath(relativeConditionRootPath, concatFilePath(type.name(), "case" + keyString +".json"));
+        String relativePath = concatFilePath(relativeConditionRootPath, concatFilePath(type.getName(), "case" + keyString +".json"));
         String filePath = concatFilePath(projectRootPath, relativePath);
         return filePath;
     }
 
     // 待生成/写入的文件
     public String getResultFilePath(String keyString, NetworkType type) {
-        String relativePath = concatFilePath(relativeLocalizeResultRootPath, concatFilePath(type.name(), "case" + keyString +".json"));
+        String relativePath = concatFilePath(relativeLocalizeResultRootPath, concatFilePath(type.getName(), "case" + keyString +".json"));
         String filePath = concatFilePath(projectRootPath, relativePath);
         return filePath;
     }
 
     // 待生成/写入的文件
     public String getPreResultFilePath(String keyString, NetworkType type) {
-        String relativePath = concatFilePath(relativeLocalizeResultRootPath, concatFilePath(type.name(), "(pre)case" + keyString +".json"));
+        String relativePath = concatFilePath(relativeLocalizeResultRootPath, concatFilePath(type.getName(), "(pre)case" + keyString +".json"));
         String filePath = concatFilePath(projectRootPath, relativePath);
         return filePath;
     }
 
     // 待读取的目录
     public static String getCfgRootPath(String keyString, NetworkType type) {
-        String bgpProvRootPath = concatFilePath(projectRootPath, relativeConfigRootPath + type.name());
+        String bgpProvRootPath = concatFilePath(projectRootPath, relativeConfigRootPath + type.getName());
         String finalPath = concatFilePath(bgpProvRootPath, "case"+keyString);
         return finalPath;
     }
 
     // 待读取的文件
     public String getViolateRulePath(String keyString, NetworkType type) {
-        String relativePath = concatFilePath(relativeVioRuleRootPath, concatFilePath(type.name(), "ViolatedRules_Case" + keyString +".json"));
+        String relativePath = concatFilePath(relativeVioRuleRootPath, concatFilePath(type.getName(), "ViolatedRules_Case" + keyString +".json"));
         String filePath = concatFilePath(projectRootPath, relativePath);
         return filterInvalidFilePath(filePath);
     }
 
     // 待读取的文件
     public static String getErrorProvFilePath(String keyString, NetworkType type, String fileName) {
-        String ProvRootPath = concatFilePath(projectRootPath, relativeErrProvRootPath + type.name());
+        String ProvRootPath = concatFilePath(projectRootPath, relativeErrProvRootPath + type.getName());
         String finalPath = concatFilePath(ProvRootPath, concatFilePath("case"+keyString, fileName));
         return filterInvalidFilePath(finalPath);
     }
 
     // 待读取的文件
     public static String getSseProvFilePath(String keyString, NetworkType type, String fileName) {
-        String ProvRootPath = concatFilePath(projectRootPath, relativeSseProvRootPath + type.name());
+        String ProvRootPath = concatFilePath(projectRootPath, relativeSseProvRootPath + type.getName());
         String finalPath = concatFilePath(ProvRootPath, concatFilePath("case"+keyString, fileName));
         return filterInvalidFilePath(finalPath);
     }
 
     // 待读取的文件
     public static String getCorrectProvFilePath(String keyString, NetworkType type, String fileName) {
-        String ProvRootPath = concatFilePath(projectRootPath, relativeErrProvRootPath + type.name());
+        String ProvRootPath = concatFilePath(projectRootPath, relativeErrProvRootPath + type.getName());
         String finalPath = concatFilePath(ProvRootPath, concatFilePath("case"+keyString, concatFilePath(KeyWord.CORRECT, fileName)));
         return filterInvalidFilePath(finalPath);
     }
 
     // 待读取的文件
     public String getPeerInfoPath(String keyString, NetworkType type) {
-        String relativePath = concatFilePath(relativePeerInfoRootPath, concatFilePath(type.name(), "PeerInfo" + keyString +".json"));
+        String relativePath = concatFilePath(relativePeerInfoRootPath, concatFilePath(type.getName(), "PeerInfo" + keyString +".json"));
         String filePath = concatFilePath(projectRootPath, relativePath);
         return filterInvalidFilePath(filePath);
     }
 
     public String getErrorDstName(String keyString, NetworkType type) {
-        if (type.equals(NetworkType.IPMETRO)) {
-            return errIpmetroDstNameMap.get(keyString);
-        } else if (type.equals(NetworkType.IPRAN)){
-            return errIpranDstNameMap.get(keyString);
-        } else if (type.equals(NetworkType.CLOUDNET)) {
-            return "";
+        switch (type) {
+            case IPMETRO:{
+                return errIpmetroDstNameMap.get(keyString);
+            }
+            case IPRAN:{
+                return errIpranDstNameMap.get(keyString);
+            }
+            case IPRAN_NEW:{
+                return errIpranNewDstName1;
+            }
+            case CLOUDNET:{
+                return errCloudnetDstName1;
+            }
+            case IPMETRO_NEW:{
+                if (ipmetroNewCaseType1.contains(keyString) || ipmetroNewCaseType3.contains(keyString)) {
+                    return errIpmetroNewDstName1;
+                }
+                if (ipmetroNewCaseType2.contains(keyString)) {
+                    return errIpmetroNewDstName2;
+                }
+            }
         }
+
         return "";
     }
 
     public String getErrorVpnName(NetworkType type) {
-        if (type.equals(NetworkType.IPMETRO)) {
+        if (type.equals(NetworkType.IPMETRO) || type.equals(NetworkType.IPMETRO_NEW)) {
             return KeyWord.PUBLIC_VPN_NAME;
-        } else if (type.equals(NetworkType.IPRAN)){
+        } else if (type.equals(NetworkType.IPRAN) || type.equals(NetworkType.IPRAN_NEW)){
             return "LTE_RAN";
         } else if (type.equals(NetworkType.CLOUDNET)) {
             return "YiLiao";
@@ -236,13 +324,32 @@ public class InputData {
     }
 
     public String getErrorDstIp(String keyString, NetworkType type) {
-        if (type.equals(NetworkType.IPMETRO)) {
-            return errIpmetroDstIpMap.get(keyString);
-        } else if (type.equals(NetworkType.IPRAN)){
-            return errIpranDstIpMap.get(keyString);
-        } else if (type.equals(NetworkType.CLOUDNET)) {
-            return "";
+        switch (type) {
+            case IPMETRO:{
+                return errIpmetroDstIpMap.get(keyString);
+            }
+            case IPRAN:{
+                return errIpranDstIpMap.get(keyString);
+            }
+            case IPRAN_NEW:{
+                return errIpranNewDstIp1;
+            }
+            case CLOUDNET:{
+                return errCloudnetDstIp1;
+            }
+            case IPMETRO_NEW:{
+                if (ipmetroNewCaseType1.contains(keyString)) {
+                    return errIpmetroNewDstIp1;
+                }
+                if (ipmetroNewCaseType2.contains(keyString)) {
+                    return errIpmetroNewDstIp2;
+                }
+                if (ipmetroNewCaseType3.contains(keyString)) {
+                    return errIpmetroNewDstIp3;
+                }
+            }
         }
+
         return "";
     }
 
@@ -252,7 +359,7 @@ public class InputData {
         } else if (type.equals(NetworkType.IPRAN)){
             return "";
         } else if (type.equals(NetworkType.CLOUDNET)) {
-            return "";
+            return corCloudnetDstName1;
         }
         return "";
     }
@@ -263,7 +370,7 @@ public class InputData {
         } else if (type.equals(NetworkType.IPRAN)){
             return "";
         } else if (type.equals(NetworkType.CLOUDNET)) {
-            return "";
+            return corCloudnetDstIp1;
         }
         return "";
     }
