@@ -14,7 +14,6 @@ import org.sng.main.diagnosis.Generator;
 import org.sng.main.util.ConfigTaint;
 import org.sng.main.util.KeyWord;
 
-import com.google.common.net.PercentEscaper;
 
 /*
  * Localize "violateIbgpPeer"/"violateEbgpPeer" errors 
@@ -79,10 +78,10 @@ public class PeerLocalizer implements Localizer{
             // 两边都配过peer, 是不一致的问题
             if (!localPeer.isConsistent(remotePeer)) {
                 // ip或者as-num不一致, 至少有一个错了, 顺着诊断一遍
-                if (!localPeer.getLocalIp().equals(remotePeer.getPeerIp())) {
+                if (!localPeer.getLocalIpString().equals(remotePeer.getPeerIpString())) {
                     errList.add(PeerErrorType.PEER_IP_INCONSISTENT_LOCAL);
                 } 
-                if (!remotePeer.getLocalIp().equals(localPeer.getPeerIp())) {
+                if (!remotePeer.getLocalIpString().equals(localPeer.getPeerIpString())) {
                     errList.add(PeerErrorType.PEER_IP_INCONSISTENT_REMOTE);
                 }
                 if (localPeer.getLocalAsNum()!=remotePeer.getPeerAsNum()) {
@@ -104,7 +103,7 @@ public class PeerLocalizer implements Localizer{
                 }
                 if (isLocalConnectInterface && !isLocalIgnorePeer) {
                     if (localPeer.getBgpPeerType()==BgpPeerType.EBGP) {
-                        int atLeastHop = generator.hopNumberToReachIpUsingStatic(localNode, localPeer.getPeerIp());
+                        int atLeastHop = generator.hopNumberToReachIpUsingStatic(localNode, localPeer.getPeerIpString());
                         if (atLeastHop==0) {
                             errList.add(PeerErrorType.PEER_IP_REACH_LOCAL);
                         } else if (atLeastHop > localPeer.getEBgpMaxHop()) {
@@ -125,7 +124,7 @@ public class PeerLocalizer implements Localizer{
                 }
                 if (isRemoteConnectInterface && !isRemoteIgnorePeer) {
                     if (remotePeer.getBgpPeerType()==BgpPeerType.EBGP) {
-                        int atLeastHop = generator.hopNumberToReachIpUsingStatic(remoteNode, remotePeer.getPeerIp());
+                        int atLeastHop = generator.hopNumberToReachIpUsingStatic(remoteNode, remotePeer.getPeerIpString());
                         if (atLeastHop==0) {
                             errList.add(PeerErrorType.PEER_IP_REACH_REMOTE);
                         } else if (atLeastHop > remotePeer.getEBgpMaxHop()) {
@@ -152,23 +151,23 @@ public class PeerLocalizer implements Localizer{
         for (PeerErrorType err : errorTypes) {
             switch (err) {
                 case PEER_AS_NUMBER_INCONSISTENT_LOCAL: {
-                    String[] keyWords = {"peer", localPeer.getPeerIp().toString(), String.valueOf(localPeer.getPeerAsNum())};
+                    String[] keyWords = {"peer", localPeer.getPeerIpString().toString(), String.valueOf(localPeer.getPeerAsNum())};
                     lines.putAll(ConfigTaint.peerTaint(localNode, keyWords));
                     break;
                 }
                 // case PEER_AS_NUMBER_INCONSISTENT_REMOTE: {
-                //     String[] keyWords = {"peer", remotePeer.getPeerIp().toString(), String.valueOf(remotePeer.getPeerAsNum())};
+                //     String[] keyWords = {"peer", remotePeer.getPeerIpString().toString(), String.valueOf(remotePeer.getPeerAsNum())};
                 //     lines.putAll(ConfigTaint.taint(remoteNode, keyWords));
                 //     break;
                 // }
                 case PEER_CONNECT_INTERFACE_LOCAL: {
                     // 这个错误默认缺失对应语句
-                    String line = "peer " + localPeer.getPeerIp().toString() + " connect-interface " + localPeer.getLocalIp().toString();
+                    String line = "peer " + localPeer.getPeerIpString().toString() + " connect-interface " + localPeer.getLocalIpString().toString();
                     lines.put(violation.getMissingLine(), line);
                     break;
                 }
                 // case PEER_CONNECT_INTERFACE_REMOTE: {
-                //     String line = "peer " + remotePeer.getPeerIp().toString() + " connect-interface " + remotePeer.getLocalIp().toString();
+                //     String line = "peer " + remotePeer.getPeerIpString().toString() + " connect-interface " + remotePeer.getLocalIpString().toString();
                 //     lines.put(violation.getMissingLine(), line);
                 //     break;
                 // }
@@ -180,54 +179,54 @@ public class PeerLocalizer implements Localizer{
                 case PEER_IP_INCONSISTENT_LOCAL:
                 case PEER_IP_REACH_LOCAL: {
                     // IP不一致 就把所有 peer *ip* 有关的语句都找出来
-                    String[] keyWords = {"peer", localPeer.getPeerIp().toString()};
+                    String[] keyWords = {"peer", localPeer.getPeerIpString().toString()};
                     lines.putAll(ConfigTaint.taint(localNode, keyWords));
                     break;
                 }
                 // case PEER_IP_INCONSISTENT_REMOTE: 
                 // case PEER_IP_REACH_REMOTE: {
-                //     String[] keyWords = {"peer", remotePeer.getPeerIp().toString()};
+                //     String[] keyWords = {"peer", remotePeer.getPeerIpString().toString()};
                 //     lines.putAll(ConfigTaint.taint(remoteNode, keyWords));
                 //     break;
                 // }
 
                 case PEER_NOT_CONFIGURED_LOCAL: {
-                    String line1 = "peer " + localPeer.getPeerIp().toString() + " enable";
+                    String line1 = "peer " + localPeer.getPeerIpString().toString() + " enable";
                     lines.put(violation.getMissingLine(), line1);
-                    String line2 = "peer " + localPeer.getPeerIp().toString() + " connect-interface " + localPeer.getLocalIp().toString();
+                    String line2 = "peer " + localPeer.getPeerIpString().toString() + " connect-interface " + localPeer.getLocalIpString().toString();
                     lines.put(violation.getMissingLine(), line2);
                     break;
                 }
                 // case PEER_NOT_CONFIGURED_REMOTE: {
-                //     String line1 = "peer " + remotePeer.getPeerIp().toString() + " enable";
+                //     String line1 = "peer " + remotePeer.getPeerIpString().toString() + " enable";
                 //     lines.put(violation.getMissingLine(), line1);
-                //     String line2 = "peer " + remotePeer.getPeerIp().toString() + " connect-interface " + remotePeer.getLocalIp().toString();
+                //     String line2 = "peer " + remotePeer.getPeerIpString().toString() + " connect-interface " + remotePeer.getLocalIpString().toString();
                 //     lines.put(violation.getMissingLine(), line2);
                 //     break;
                 // }
                 case EBGP_MAX_HOP_LOCAL: {
-                    int realHop = generator.hopNumberToReachIpUsingStatic(localNode, localPeer.getPeerIp());
-                    String line = "peer " + localPeer.getPeerIp().toString() + " ebgp-max-hop " + String.valueOf(realHop);
-                    String[] keyWords = {"peer", localPeer.getPeerIp().toString(), "ebgp-max-hop"};
+                    int realHop = generator.hopNumberToReachIpUsingStatic(localNode, localPeer.getPeerIpString());
+                    String line = "peer " + localPeer.getPeerIpString().toString() + " ebgp-max-hop " + String.valueOf(realHop);
+                    String[] keyWords = {"peer", localPeer.getPeerIpString().toString(), "ebgp-max-hop"};
                     lines.putAll(ConfigTaint.peerTaint(localNode, keyWords));
                     lines.put(violation.getMissingLine(), line);
                     break;
                 }
                 // case EBGP_MAX_HOP_REMOTE: {
-                //     int realHop = generator.hopNumberToReachIpUsingStatic(remoteNode, remotePeer.getPeerIp());
-                //     String line = "peer " + remotePeer.getPeerIp().toString() + " ebgp-max-hop " + String.valueOf(realHop);
-                //     String[] keyWords = {"peer", remotePeer.getPeerIp().toString(), "ebgp-max-hop"};
+                //     int realHop = generator.hopNumberToReachIpUsingStatic(remoteNode, remotePeer.getPeerIpString());
+                //     String line = "peer " + remotePeer.getPeerIpString().toString() + " ebgp-max-hop " + String.valueOf(realHop);
+                //     String[] keyWords = {"peer", remotePeer.getPeerIpString().toString(), "ebgp-max-hop"};
                 //     lines.putAll(ConfigTaint.taint(remoteNode, keyWords));
                 //     lines.put(violation.getMissingLine(), line);
                 //     break;
                 // }
                 case UNKOWN_LOCAL: {
-                    String[] keyWords = {"peer", localPeer.getPeerIp().toString()};
+                    String[] keyWords = {"peer", localPeer.getPeerIpString().toString()};
                     lines.putAll(ConfigTaint.peerTaint(localNode, keyWords));
                     break;
                 }
                 // case UNKOWN_REMOTE: {
-                //     String[] keyWords = {"peer", remotePeer.getPeerIp().toString()};
+                //     String[] keyWords = {"peer", remotePeer.getPeerIpString().toString()};
                 //     lines.putAll(ConfigTaint.taint(remoteNode, keyWords));
                 //     break;
                 // }
@@ -239,10 +238,10 @@ public class PeerLocalizer implements Localizer{
     private boolean isConnectInterface(String node) {
         // node上对 对端的peer配置 是否有connect-interface命令
         if (node.equals(localNode)) {
-            String[] keyWords = {"peer", localPeer.getPeerIp().toString(), "connect-interface"};
+            String[] keyWords = {"peer", localPeer.getPeerIpString().toString(), "connect-interface"};
             return ConfigTaint.taint(node, keyWords).keySet().size()>0;
         } else if (node.equals(remoteNode)) {
-            String[] keyWords = {"peer", remotePeer.getPeerIp().toString(), "connect-interface"};
+            String[] keyWords = {"peer", remotePeer.getPeerIpString().toString(), "connect-interface"};
             return ConfigTaint.taint(node, keyWords).keySet().size()>0;
         } else {
             return false;
@@ -254,7 +253,7 @@ public class PeerLocalizer implements Localizer{
 
     private boolean isIgnorePeer(String node) {
         // node上对 对端的peer配置 是否有peer ignore命令
-        String[] keyWords = {"peer", localPeer.getPeerIp().toString(), "ignore"};
+        String[] keyWords = {"peer", localPeer.getPeerIpString().toString(), "ignore"};
         if (ConfigTaint.taint(node, keyWords).keySet().size()>0) {
             return true;
         }
